Factor out agent code from executables
